third_party/WebKit/LayoutTests/css3/compositing/blend-mode-property-expected.txt

Test the parsing of the mix-blend-mode property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be normal (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be multiply (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be screen (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be overlay (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be darken (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be lighten (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be color-dodge (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be color-burn (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be hard-light (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be soft-light (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be difference (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be exclusion (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be hue (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be saturation (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be color (of type string). Was null (of type object).
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('mix-blend-mode') should be luminosity (of type string). Was null (of type object).
PASS successfullyParsed is true

TEST COMPLETE
